Move the entire snaphu/spurt redo conncomp functions to utils
