+++ src/disp_s1/main.py
@@ -13,7 +13,7 @@ from dolphin import PathOrStr, io, stitching
 from dolphin._log import log_runtime, setup_logging
 from dolphin.unwrap._utils import create_combined_mask
 from dolphin.utils import DummyProcessPoolExecutor, full_suffix, get_max_memory_usage
-from dolphin.workflows.config import DisplacementWorkflow, UnwrapOptions
+from dolphin.workflows.config import DisplacementWorkflow
 from dolphin.workflows.displacement import OutputPaths
 from dolphin.workflows.displacement import run as run_displacement
 from opera_utils import get_dates, group_by_burst, group_by_date
@@ -25,7 +25,11 @@ from disp_s1._ps import precompute_ps
 from disp_s1.pge_runconfig import RunConfig
 
 from ._reference import ReferencePoint, read_reference_point
-from ._utils import _create_correlation_images, _regrow
+from ._utils import (
+    _create_correlation_images,
+    _update_snaphu_conncomps,
+    _update_spurt_conncomps,
+)
 
 logger = logging.getLogger(__name__)
 
@@ -515,78 +519,6 @@ def create_displacement_products(
                 repeat(near_far_incidence_angles),
             )
         )
-
-
-def _update_snaphu_conncomps(
-    timeseries_paths: Sequence[Path],
-    stitched_cor_paths: Sequence[Path],
-    mask_filename: PathOrStr,
-    unwrap_options: UnwrapOptions,
-    nlooks: int,
-    max_workers: int = 2,
-) -> list[Path]:
-    """Recompute connected components from SNAPHU after a timeseries inversion.
-
-    `timeseries_paths` contains the post-inversion rasters, one per secondary date.
-
-    Parameters
-    ----------
-    timeseries_paths : list[Path]
-        list of paths to the timeseries files.
-    stitched_cor_paths : list[Path]
-        list of paths to the pseuedo-correlation rasters.
-    mask_filename : PathOrStr
-        Path to a binary mask matching shape of `timeseries_paths`.
-    unwrap_options : [dolphin.workflows.config.UnwrapOptions][]
-        Configuration object containing unwrapping options.
-    nlooks : int
-        Effective number of looks used to make correlation.
-    max_workers : int
-        Number of parallel files to process.
-        Default is 2.
-
-    Returns
-    -------
-    list[Path]
-        list of updated connected component paths.
-
-    """
-    args_list = [
-        (unw_f, cor_f, nlooks, mask_filename, unwrap_options)
-        for unw_f, cor_f in zip(timeseries_paths, stitched_cor_paths)
-    ]
-
-    mp_context = get_context("spawn")
-    with mp_context.Pool(max_workers) as pool:
-        return list(pool.map(_regrow, args_list))
-
-
-def _update_spurt_conncomps(
-    timeseries_paths: Sequence[Path],
-    conncomp_paths: Sequence[Path],
-) -> list[Path]:
-    """Recompute connected components from spurt after a timeseries inversion.
-
-    Parameters
-    ----------
-    timeseries_paths : list[Path]
-        list of paths to the timeseries files.
-    conncomp_paths : list[Path]
-        list of connected component paths from the spurt unwrapping
-
-    Returns
-    -------
-    list[Path]
-        list of updated connected component paths.
-
-    """
-    new_conncomp_paths: list[Path] = []
-    for cc_p, ts_p in zip(conncomp_paths, timeseries_paths, strict=False):
-        new_name = cc_p.parent / str(ts_p.name).replace(
-            full_suffix(ts_p), full_suffix(cc_p)
-        )
-        new_conncomp_paths.append(cc_p.rename(new_name))
-    return new_conncomp_paths
 
 
 def _create_nodata_mask(filename: PathOrStr, output_filename: PathOrStr) -> None:
